+++ components/campaign/CampaignTable.js
@@ -8,7 +8,11 @@ export default function CampaignTable (props) {
   if (props.users.length === 0) {
     return <div />
   }
-
+  const campaignTopStats = sortBy(prop('edits'), props.users).reverse()
+    .map(user => ({
+      ...user,
+      km_coastlines_all: user.km_coastlines_add + user.km_coastlines_mod
+    }))
   return (
     <div className='widget clearfix table-wrapper'>
       <table>
@@ -27,8 +31,7 @@ export default function CampaignTable (props) {
         </thead>
         <tbody>
           {
-            sortBy(prop('edits'), props.users)
-              .reverse()
+            campaignTopStats
               .map((user, idx) => (
                 <tr key={user.uid}>
                   <td>{idx + 1}</td>
@@ -42,7 +45,7 @@ export default function CampaignTable (props) {
                   <td>{formatDecimal(user.km_roads_add)}</td>
                   <td>{formatDecimal(user.buildings_add)}</td>
                   <td>{formatDecimal(user.poi_add)}</td>
-                  <td>{formatDecimal(user.km_coastlines_add + user.km_coastlines_mod)}</td>
+                  <td>{formatDecimal(user.km_coastlines_all)}</td>
                   <td>{formatDecimal(user.km_waterways_add)}</td>
                   <td>{formatDecimal(user.edits)}</td>
                   <td>{formatDecimal(user.editSum)}</td>
@@ -51,12 +54,12 @@ export default function CampaignTable (props) {
           }
         </tbody>
       </table>
-      <CSVLink className='link--large' style={{ display: 'inline-block', float: 'right', marginTop: '2rem' }} data={sortBy(prop('edits'), props.users).reverse()} filename={`${props.name} - Top 10 Participants.csv`} headers={[
+      <CSVLink className='link--large' style={{ display: 'inline-block', float: 'right', marginTop: '2rem' }} data={campaignTopStats} filename={`${props.name} - Top 10 Participants.csv`} headers={[
         { label: 'Name', key: 'name' },
         { label: 'Roads (Km)', key: 'km_roads_add' },
         { label: 'Buildings', key: 'buildings_add' },
         { label: 'Points of Interest', key: 'poi_add' },
-        { label: 'Coastlines (Km)', key: 'km_coastlines_add' },
+        { label: 'Coastlines (Km)', key: 'km_coastlines_all' },
         { label: 'Waterways (Km)', key: 'km_waterways_add' },
         { label: 'Changesets', key: 'edits' },
         { label: 'Edits', key: 'editSum' }